Return `None` for invalid state in `CalcJobNode.get_state` (#3068)
The `state` attribute of a calculation job node is an active sub state
specific only to `CalcJobs` to give more granularity to the general
process state. This attribute is only set during the lifetime of the
calculation job and is removed afterwards. The docstring now makes this
clear to prevent confusion when it returns `None`. The method now also
returns `None` if the attribute contains an invalid value, instead of
excepting, which should only happen due to incorrectly migrated legacy
data. We do not want this to except the application.