Add concept of "mutable" model fields in the `ModelWrapper`
The goal of the `ModelWrapper` is to ensure that getting and setting
fields of a database model instance automatically take care of
refreshing the state from the database and flushing any changes.
This is at the cost of more database operations which should be kept to
a minimum especially if they are useless.

For example, the refreshing and flushing of fields that are immutable
model fields when the model is already stored. The get and set attribute
methods are improved to ignore the refresh and flush if the model is
already saved and the field is immutable. Currently these are hard coded
to the set of `pk`, `id` and `uuid` since these are always immutable for
all the database models implemented in AiiDA.