Make `Process.exposed_inputs/exposed_outputs` return `AttributeDict` (#3079)
The return value of these methods should be mutable. However, the source
mapping is often immutable, as it should be, as they represent process
inputs and outputs, respectively. Therefore, in any case, the result
mapping should be a deep copy and should be mutable. Simply wrapping the
result in an `AttributeDict`, which recursively wraps all nested
mappings, will satisfy both criteria.