Only print `retrieve_singlefile_list` deprecation when necessary
While the functionality still exists, the engine will keep calling the
`get_retrieve_singlefile_list` method and try to set them from the
calculation info object, so the warnings were printed even when the user
was not actually using it. Make sure that the relevant methods are only
called when the user explicitly sets the `retrieve_singlefile_list`.