Update the export archive version to `v0.6`
The change in the export archive is necessary to take the changes into
account of two database migrations:

 * Moving of node attributes and extras to JSONB for the Django backend
 * Data migration of `state` attribute of legacy `JobCalculation` nodes

Due to the migration of node attributes and extras to JSONB, which does
not support datetime objects, in contrast with the old database schema,
means that the deserialization from the export JSON data is no longer
necessary. The only serialized types were datetime objects, but these
are now no longer supported anyway in the database, just like SqlAlchemy
already was. This allows us to remove the attribute and extra conversion
dictionaries in its entirety.

A second migration deals with the conversion of legacy job calculation
states, by inferring the `process_state`, `exit_status` and
`process_status` from it.

Co-authored-by: Giovanni Pizzi <giovanni.pizzi@epfl.ch>