Force tags to zero when using ASE to convert a CIF file. (#2942)
After `ase==3.12.0`, the `ase.io.read` method for CIF files changed the
usage of "tags". Originally the tags were always zero, but in the newer
versions, the tags were incremented for each new elemental species that
occurred in the CIF file. Since the `ase` object returned by `read` is
used to construct `StructureData` objects, which reads the tag names to
build the kind names, the behavior of `StructureData` was also changed.
To restore the old behavior, we force the tags to be all zeros.