Enforcing specific precision in clean_value for floats
This will guarantee a larger uniformity in what gets stored
across backends. This fixes #2631

Note that this commit also fixes the problem of the bigint vs. float
part of the discussion that is mentioned in the thread, i.e.
that numbers like 1.2e20, floats in python, become long ints
in the database. Here we forcefully convert numbers >=1.e14 to
integers to alleviate the problem and ensure that a roundtrip
via the database does not change the hash.

IMPORTANT: in this commit, we enforce a maximum precision of 14
digits in AiiDA - this is a change with respect to what used
to be before. Note also that we are *not* going to truncate old
AiiDA values, so behavior before and after this commit will change
even if the stored value is the same.

Note that this commit changes the hashing functionality, so we
will need to ask users to rehash everything.