Allow to ignore default values in `InteractiveOption`
The `InteractiveOption` in interactive mode, will populate the value
with a default, if defined. If no value is specified by the user, this
is interpreted as the user wanting to use the default. However, since
the defaults are often contextual and are determined by existing data,
the situation can occur where one wants to set no specific value, even
though a default has been defined through pre-existing data. To fix this
situation, a special character `!` is introduced that when passed in
interactive mode is interpreted as passing `None`. Note that this only
works in interactive mode. When `!` is used non-interactively, the value
is interpreted as is.