Improve and simplify the guard of `verdi database migrate` (#3110)
The additional guards were added during the beta phase of `v1.0.0` to
make sure people were aware of the potential danger of migrating their
database. Now that the migrations have been field tested, we can reduce
the confirmation prompt to a single one. Additionally, the message
prompted when the database is out of date as well as the command to
update the database itself, have been improved to explicitly state that
any running processes have to be finished on the old installed version
first, before migrating the data.