Properly overload function module and name to `FunctionProcess`
The `FunctionProcess` class that is dynamically built based on the
function that is decorated with the `calfunction` or `workfunction`
decorator was not overloading the `__module__` and `__name__` attributes
of the wrapped function. Since the `process_type` attribute of the
process node is defined by those attributes, an incorrect basic value
was being stored of the form `abc.name_of_wrapped_function`.