Ensure immutability of `CalcJobNode` hash before and after storing
The `Node._get_objects_to_hash` method, which returns all the objects of
a `Node` instance that should be included in computing its hash, also
included the file repository. This proved problematic for the
`CalcJobNode` sub class. The hash of a node is computed during the store
method, in order to determine if it can be cached from an existing node
with an identical hash. However, at that point, the repository of the
node is empty, but as soon as the node is stored and process is started,
the input files that are generated by the `CalcJob` class based on the
inputs, are stored in the repository of the node. If the hash were to be
recomputed at that point, it would be different from the original hash
computed during storing. This breaks the caching mechanism.

Since the input files that are written to the repository by the `CalcJob`
are derivatives of the input nodes (which are included in the hash), and
therefore do not semantically add anything to the provenance, we can
simply ignore them. Since this only applies to the `CalcJobNode`, we
simply override the method in that sub class and omit the repository.

N.B.: in this commit we also loosen the condition of when a completed
process node is considered to be a valid cache. Before only processes
that finished successfully (i.e. with process state `finished` and an
exit status `0`) were considered valid caches. This is loosened to also
accept non-zero exit statuses. The new rule then considers all processes
that have `finished` as a valid cache, excluding only `excepted` and
`killed` processes.