Add data migration for legacy process attributes
Attribute keys that are renamed:

  * `_sealed` -> `sealed`

Attribute keys that are removed entirely:

  * `_finished`
  * `_failed`
  * `_aborted`
  * `_do_abort`

The `sealed` key is still used, the key has just changed, so existing
attributes are migrated. The last four attributes are left over control
attributes of one of the first process based engines, but are obsolete
and are therefore deleted to clean up some space.

Finally, any process nodes that still do not have a sealed attribute,
get one, because these most likely correspond to legacy calculation
nodes that were run before the sealed concept existed. However, any
nodes that have a `process_state` that is active, are excluded as they
correspond to actual active processes and are therefore not yet sealed.