Make the `execmanager.upload_calculation` idempotent'ish
The `upload_calculation` would cause an exception if called multiple
times for the same calculation, which can happen if the first time that
the runner was working on it got interrupted, for example due to a
daemon shutdown. The reason is that the second time around the adding of
the `remote_folder` data node will raise a uniqueness exception,
because there can only be one output with the same label.

Note that full idem-potency is impossible, but this change should make
the problem a lot less likely to occur. The idea is to delay the actual
attaching of the remote folder data node to the last moment possible.
This way, if the method is called again and the folder is already there,
we can be reasonably sure that the files were already retrieved
successfully and we simply return, leaving the call a no-op. This is
done in the beginning of the function to check if the output node already
exists using the `LinkManager.first()` call. If the node exists, the
upload function has apparently already been called before and reached
the end of the function where it adds the remote folder. This means all
the files were already successfully uploaded so we can safely skip it.