Fix bug with never resolving calculation job update futures
The `JobsList` class is a container and manager of requests for updates
of the status of calculation jobs. For each request, it keeps a future
that it will update in bulk after retrieving the status of all the jobs
that it manages. This is necessary to ensure that multiple jobs use the
same update call to scheduler and not all individually, in order to not
overload the scheduler.

If the scheduler update call excepts, all the futures were properly
excepted as well and the internal cache of outstanding requests would be
cleared, before re-raising the exception. The excepted futures would
each be caught by the exponential backoff mechanism and a new request to
update the status would be scheduled. This would be registered with the
same `JobsList` instance, however, the next scheduler update would never
be called and so the futures would never be resolved, causing the
calculation jobs to "hang" indefinitely.

The problem lies in the `_ensure_updating` method. When a new request
comes in, this is the method that is called first to see if a scheduler
update call has already been scheduled. If the case, a new one does not
have to be scheduled. This is based on whether the `_update_handle`
attribute is defined. This attribute is properly reset after a scheduler
update goes through successfully, in the `updating` coroutine after the
`_update_job_info` yields. However, in the case of an exception in that
last method, the code path that resets the `_update_handle` is never
reached because the exception is re-raised after the job requests cache
is reset. The solution is simply to manually set `_update_handle` to
`None` in the except clause.