Docs: fix validator description, add warning for `ToContext`
The signature description of the `Port.validator` incorrectly stated
that it should return a boolean. In reality it should return nothing for
valid values and return a string with the validation error message for
invalid values.

In addition, a warning box has been added to the description of the
`ToContext` construct for `WorkChains` explaining that adding awaitables
through this mechanism will cause the `WorkChain` to wait for them to be
completed, if and only if there is a next step in the outline. If the
addition of the awaitables is done in the last step of the outline, the
engine will consider the work of the `WorkChain` as completed and will
terminate the process.