Simplify the link triple uniqueness determination  in `validate_link`
In the case of a `unique_triple` degree, the link triple should be
unique. This means that to check this, we just have to find a single one
that already exists to know that the condition is validated. It is not
expected to reduce the computation time by much, but the logic is a lot
clearer.