Docs: updating "Transport plugins" for `aiida-core==1.0.0`
The inline references of the base class and the local and ssh plugins
have been replaced with a reference to the automatically generated API
references. This interrupts the flow of the text less. The template for
a transport plugin is now provided as a downloadable link instead of
being copied inline.