Minor corrections of documentation sections: REST API and result manager; typos in AUTHORS
Minor correction to make sure the documentation regarding REST API and result manager
are up to date with the latest version of AiiDA 1.0.

Dealing with issues #3048 and #3027

Checked that the files are where the user is told they are, and commands work (python 3 corrections too).

Mainly corrected typos, unified  and broken links.

I checked that the examples in the REST API section work, but I did not try to
set ap an APACHE server -- I assume that has been tested by someone else elesewhere (#3106 ? )

Corrected a couple imprecision in the AUTHOR list file (double entry and an incorrect name)