Resetting migrations for Django and SqlAlchemy backend
Here we drop all existing migrations and regenerate the initial schema.
Commit `6c1391f9f7de3b01d1bfb4b2e72542fa24854195` prepared the reset of
the schema generation by storing the `schema_generation` in the
`DbSettings` table for all backends. The corresponding final database
schema version marks the point where the schema will be identical to the
brand new initial schema introduced in this commit. If an existing
database is at that transition version, the schema version and
generation can simply be changed to the new one knowing that the rest of
the state of the database will be identical. This allows us to remove
all 40 odd existing migrations that have accumulated since the beginning
of development, reducing development and testing overhead.