Prepare the reset of the database schema
Since the original databsae schema a lot changes have been applied, for
which migrations have been added to the codebase to automatically update
existing databases. However, these migrations require keeping old code
around just to keep them working as well as the for unittests, adding
significant overhead in test run time and development cost. For the
upcoming release of `aiida-core==1.0.0` we intend to "reset" the
database schema so the current migrations can be removed. To do this, we
add a final migration that will mark the point at which the reset can
take place. For convenience we add a setting `schema_generation` which
indicates the current schema generation, which for the `0.*` versions of
`aiida-core` is set to generation `1`.