Add tests for `dry_run=True` and `store_provenance=False` (#3202)
These tests ensure that the various launchers respect the return value
signature irrespective of the values of the `dry_run` and
`store_provenance` metadata inputs. Also they add an explicit test for
the presence of the `dry_run_info` attribute on the returned node which
was added in a recent commit.