Refactor the code dealing with database into `BackendManager`
The code that acts directly with the database, outside of the ORM, for
example to check the database schema generation and version, or to get
data from the `DbSettings` table has been refactored. The ORM requires
the database environment to be loaded, but to do that the database
schema needs to be verified which also requires accessing the database.
This functionality was implemented mostly in free-standing functions and
separately for both backends. This is reorganized into a `BackendManager`
abstract class and implemented for Django and SqlAlchemy. This
centralizes the functionality required to verify database schema and
perform the migrations thereof.