Adding documentation on how to get the AuthInfo
Fixes #3195

Adds a cookbook recipe on how to get an AuthInfo
object given, e.g., a computer name, and how to then
get a Transport instance. This is not what a user would
typically do, but it is still very often useful for
intermediate users who want to e.g. perform operations
on the computer via the transport.