Make sure `verdi` commands respect PEP 257 docstring conventions
The PEP states that a docstring should start with a single short
description ending with a full stop. The `click` library will take this
as the short help message when displaying the help message of the parent
group. When the message is also sufficienty short, no ellipsis will be
displayed.

P.S.: the "easter-egg" is now a proper hidden easter egg.