Docs: caching implementation and database schema
The page explaining the database schema was horribly outdated and not
very useful. It essentially contained a single automatically generated
schema from the Django database. The page has been removed.

The page with details on the caching implementation has been slightly
adapted to be more clear, but the content was still correct.

The `_cachable` attribute, which was already set to `True` in the
`CalculationNode` class, was set again in the `CalcJobNode` sub class,
which is superfluous and has therefore been removed.