Fix bugs in `verdi process watch`
The signature of the callback function did not define the `communicator`
causing an exception to be thrown. The callback function is improved to
make the formatted messages easier to read. The main command also has
extra information message echoed to explain to the user how to interrupt
the command. Finally, the usage of a `Future` to block the thread
indefinitely has been replaced by a simple while loop calling `sleep`
because the future was swallowing the keyboard interrupt, making it
impossible to exit from the command that way.