Fix eagerness of `verdi setup --help` being overruled (#3257)
The eagerness of the `--help` option for the `verdi setup` and `verdi
quicksetup` commands was being overruled by the `--profile` option. This
is because the latter was also made eager, such that `verdi setup --help`
actually prompted for the profile name instead of showing the help.

The profile option was made eager in an attempt to solve another problem
where some of the other options of the command have contextual defaults
that depend on the value of the profile. It therefore needs to be parsed
first. In `click` the parse order is determined by the order in which
the options are defined on the command, regardless of the order in which
they are passed by the user. This was done correctly, but still the
contextual default of, for example the `--repository` option, was called
before the `--profile` was parsed, if the former was passed first. This,
however, was due to the fact that for the repository option, both the
`callback` as well as the `contextual_default` were defined, where the
former was inherited from the base option for the `verdi quicksetup`
command.

The final solution then is to remove the `callback` from the `--profile`
option for `verdi setup` and only have the `contextual_default`. This
will make sure that `--profile` will be parsed before the prompt loop
will be entered and the `contextual_default` will be called.