update instructions for generating ssh keys
AiiDA relies on paramiko for SSH connections, which - so far - only
works with SSH keys in the older "PEM" format.

Recently, the type of key generated on MacOS with
```
ssh-keygen -t rsa
```
changed from PEM to openssh, and the
[corresponding keys no longer work with paramiko](https://github.com/paramiko/paramiko/issues/340#issuecomment-492448662)
(what is doubly confusing is that key will work just fine if you use
`ssh` directly).

Changing the line in the AiiDA documentation from
```
ssh-keygen -t rsa
```

to
```
ssh-keygen -t rsa -b 4096 -m PEM
```
solves the issue.