Update instructions for generating ssh keys (#3281)
AiiDA relies on paramiko for SSH connections, which - so far - only
works with SSH keys in the older "PEM" format. Recently, the type of key
generated on MacOS with:

    ssh-keygen -t rsa

changed from PEM to openssh, and the corresponding keys no longer work
with `paramiko`. See the issue for more details:

   https://github.com/paramiko/paramiko/issues/340

what is doubly confusing is that key will work just fine if you use
`ssh` directly. Generating the key with the `-m` flag:

    ssh-keygen -t rsa -b 4096 -m PEM

restores paramiko compatibility and solves the issue.