Deprecate `mode` and `encoding` of `Repository.put_object_from_file`
When copying a file to the repository from a filepath, it does not make
sense to be able to specify the `mode` and `encoding` in which the
target file handle should be opened. It should simply copy over the
bytes and therefore always open in binary mode.

To be able to detect if the arguments are being used, we have to change
the defaults to `None` in order to be able to emit the deprecation
warning. This is justified because despite changing the defaults the
intended side-effect of the method, copying over a file, will be
unchanged in its behavior.