Rework `extract_tree` and unpacking archives
The extraction of export archives is now handled by the `Archive` class,
which will point to a `SandboxFolder` if the archive is not a folder, in
which case it will point to the folder, making it a `Folder` object
instead - as to not erase the archive when done with it.

The use of the extraction functions now solely fall upon the `Archive`
class.