Move getting completed job accounting to `retrieve` transport task
The task of retrieving the detailed accounting info from the scheduler
for completed jobs was erroneously placed within the scheduler status
update cycle of the `JobsList`. This would have requested the detailed
job info each update cycle where it not for a conditional that checked
that the scheduler status was `DONE`. However, within the `JobsList`
loop, completed jobs would not even appear in the result and hence the
conditional would never be hit. This caused the detailed accounting
never to be retrieved.

A logically better place for this functionality is in the `retrieve`
transport task after the `UPDATE` task, i.e. when the scheduler reports
that the job has terminated. The job accounting will query the scheduler
for the full accounting which, when implemented for the given scheduler
plugin, will be stored as an attribute on the calculation job node.

The `Scheduler.get_detailed_jobinfo` method is deprecated in favour of
the `Scheduler.get_detailed_job_info`. The former was returning a string
with the return value, stdout and stderr formatted within it, which is
not very flexible. The replacing method returns a dictionary.