Remove obsolete `CalcJob.get_hash` (#3318)
This overridden method was necessary originally to add the contents of
the repository to the exclude list when computing the hash. However,
recently in commit 6d4f1ecd8388500c3a0a73767cd32ecc25c25956 this was
added to the base functionality so the overridden behavior is now
identical allowing the method to be removed.